Splitting up coverage tests class
